Fix of unstarting first 2sec.
This is because debug build takes 2 times for full screen convert.
But in this case, user can operate as if it already recording started.
First we move full encode initialization to UI thread because we need to wait first muxer start before doing something (even for stop!).

In realease build, it takes about 100msec and not noticiable to GalaxyNote3.
But I'll optimize a little more later.